Update tests_admin.py and tests_admin_e2e.py
